+++ core/src/plugins/kubernetes/nginx/ingress-controller.ts
@@ -9,7 +9,6 @@
 import type { Log } from "../../../logger/log-entry.js"
 import type { KubernetesPluginContext } from "../config.js"
 import { KubeApi } from "../api.js"
-import { kindNginxInstall, kindNginxStatus, kindNginxUninstall } from "../local/kind.js"
 import { apply } from "../kubectl.js"
 import { defaultBackendStatus } from "./default-backend.js"
 import { getGenericNginxHelmValues } from "./nginx-helm-generic.js"
@@ -17,6 +16,7 @@ import { helmNginxInstall, helmNginxStatus, helmNginxUninstall } from "./nginx-h
 import { getK3sNginxHelmValues } from "./nginx-helm-k3s.js"
 import { microk8sNginxInstall, microk8sNginxStatus, microk8sNginxUninstall } from "./nginx-microk8s.js"
 import { minikubeNginxInstall, minikubeNginxStatus, minikubeNginxUninstall } from "./nginx-minikube.js"
+import { kindNginxInstall, kindNginxStatus, kindNginxUninstall } from "./nginx-kind.js"
 
 export async function ingressControllerReady(ctx: KubernetesPluginContext, log: Log): Promise<boolean> {
   const clusterType = ctx.provider.config.clusterType
